requirements.txt
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
et-xmlfile==1.1.0
exceptiongroup==1.2.1
h11==0.14.0
idna==3.7
image==1.5.33
numpy==1.26.4
openpyxl==3.1.3
outcome==1.3.0.post0
pandas==2.2.2
pillow==10.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.1
wsproto==1.2.0
